[breaking] rework of whole config file structure
- all custom commands are now stored in their own file under the plugins directory, with their respective keymaps included. dependencies for these commands are listed at the top of the config file. all commands have a short description that explains what they do
- lfrc will now only link to all these plugin config files, with the option to uncomment or delete the link, so the command can't be used
- the keymaps file will now only list the keybinds for basic lf usage

explanation: this will make everything way more flexible and easier to understand for the user. before this change all commands where included in the lfrc, which was pretty overwhelming to go through, since it had more than 1000 lines of code. now you can pick and choose which commands are useful to you and add or remove them however you want. this should also be easier to maintain and troubleshoot.